+++ ChangeLog.md
@@ -3,3 +3,9 @@
 ## 0.1.9.2  -- 2016-07-04
 
 * Tested to comform to http://github.nrel.gov:kgruchal/cesds/commit/de659ed7be4c23223a661e6d60d27a4aa9045ee1.
+
+## 0.2.6.0  -- 2016-07-04
+
+* Tested to comform to http://github.nrel.gov:kgruchal/cesds/commit/ea52ec71c809996206df5657fd0f0e3d931f25a5.
+* The delete request for work, bookmarks, and filters is not supported.
+* Retrieving only certain variables in records is not supported.